P565P14-19 : Added cumulative ratings for place

+++ GoBike/src/main/java/com/iu/gobike/dto/RatingRequest.java
@@ -14,6 +14,6 @@ public class RatingRequest {
 
     private String userName;
     private String place;
-    private Integer rating;
+    private float rating;
 
 }
